Fix tests, clean naming
