+++ src/config.rs
@@ -1,8 +1,4 @@
-use crate::{
-    lie::Lie,
-    tale::{Fib, Tale},
-    MendaxError,
-};
+use crate::{fib::Fib, MendaxError};
 use rhai::{
     Array, CustomType, Dynamic, Engine, EvalAltResult, FnPtr, Map, NativeCallContext, Scope,
     TypeBuilder,
@@ -21,13 +17,13 @@ pub fn read<P: AsRef<Path>>(fname: P, unrestricted: bool) -> Result<Lie, Box<Eva
     let engine = engine(unrestricted);
 
     let mut scope = Scope::new();
-    scope.push("lie", SharedLieBuilder::new(unrestricted));
+    scope.push("lie", SharedLie::new(unrestricted));
 
     let src = get_src(fname)?;
     let ast = engine.compile_with_scope(&scope, src)?;
     engine.run_ast_with_scope(&mut scope, &ast)?;
 
-    scope.get_value::<SharedLieBuilder>("lie").unwrap().build()
+    scope.get_value::<SharedLie>("lie").unwrap().try_into()
 }
 
 fn get_src(fname: &Path) -> Result<String, Box<EvalAltResult>> {
@@ -55,7 +51,7 @@ fn get_src(fname: &Path) -> Result<String, Box<EvalAltResult>> {
 
 fn engine(unrestricted: bool) -> Engine {
     let mut engine = Engine::new();
-    engine.build_type::<SharedLieBuilder>();
+    engine.build_type::<SharedLie>();
 
     if !unrestricted {
         engine.set_max_array_size(1000);
@@ -73,21 +69,19 @@ fn engine(unrestricted: bool) -> Engine {
 }
 
 #[derive(Clone, Debug)]
-struct SharedLieBuilder(Rc<RefCell<LieBuilder>>);
+struct SharedLie(Rc<RefCell<Lie>>);
 
-impl From<LieBuilder> for SharedLieBuilder {
-    fn from(lie: LieBuilder) -> Self {
-        Self(Rc::new(RefCell::new(lie)))
+impl From<Lie> for SharedLie {
+    fn from(tale: Lie) -> Self {
+        Self(Rc::new(RefCell::new(tale)))
     }
 }
 
-impl SharedLieBuilder {
-    fn new(allow_system: bool) -> Self {
-        Self::from(LieBuilder::new(allow_system))
-    }
+impl TryFrom<SharedLie> for Lie {
+    type Error = Box<EvalAltResult>;
 
-    fn build(self) -> Result<Lie, Box<EvalAltResult>> {
-        Ok(self
+    fn try_from(shared: SharedLie) -> Result<Self, Self::Error> {
+        Ok(shared
             .0
             .try_borrow()
             .map_err(|e| {
@@ -96,11 +90,16 @@ impl SharedLieBuilder {
                     at: None,
                 }))
             })?
-            .clone()
-            .build())
+            .clone())
+    }
+}
+
+impl SharedLie {
+    fn new(allow_system: bool) -> Self {
+        Self::from(Lie::new(allow_system))
     }
 
-    fn lie(&self, ctx: &NativeCallContext) -> Result<Ref<'_, LieBuilder>, Box<EvalAltResult>> {
+    fn lie(&self, ctx: &NativeCallContext) -> Result<Ref<'_, Lie>, Box<EvalAltResult>> {
         self.0.try_borrow().map_err(|e| {
             Box::new(EvalAltResult::from(MendaxError::LieUnreadable {
                 error: Box::new(e),
@@ -109,10 +108,7 @@ impl SharedLieBuilder {
         })
     }
 
-    fn lie_mut(
-        &self,
-        ctx: &NativeCallContext,
-    ) -> Result<RefMut<'_, LieBuilder>, Box<EvalAltResult>> {
+    fn lie_mut(&self, ctx: &NativeCallContext) -> Result<RefMut<'_, Lie>, Box<EvalAltResult>> {
         self.0.try_borrow_mut().map_err(|e| {
             Box::new(EvalAltResult::from(MendaxError::LieUnwritable {
                 error: Box::new(e),
@@ -212,30 +208,30 @@ impl SharedLieBuilder {
 }
 
 #[derive(Clone, Debug)]
-pub struct LieBuilder {
-    tale: Tale,
+pub struct Lie {
+    fibs: Vec<Fib>,
     known_tags: Rc<RefCell<HashSet<String>>>,
     allow_system: bool,
     root: bool,
 }
 
-impl LieBuilder {
+impl Lie {
     fn new(allow_system: bool) -> Self {
         Self {
-            tale: Tale::default(),
+            fibs: Vec::new(),
             known_tags: Rc::new(RefCell::new(HashSet::new())),
             allow_system,
             root: true,
         }
     }
 
-    fn build(self) -> Lie {
-        Lie::new(self.tale)
+    pub fn into_fibs(self) -> Vec<Fib> {
+        self.fibs
     }
 
     fn child(&self) -> Self {
         Self {
-            tale: Tale::default(),
+            fibs: vec![],
             known_tags: self.known_tags.clone(),
             allow_system: self.allow_system,
             root: false,
@@ -257,20 +253,20 @@ impl LieBuilder {
 
     fn run(&mut self, cmd: &str, result: Vec<String>) {
         let cmd = cmd.into();
-        self.tale.push(Fib::Run { cmd, result });
+        self.fibs.push(Fib::Run { cmd, result });
     }
 
     fn show(&mut self, text: &str) {
         let text = text.into();
-        self.tale.push(Fib::Show { text });
+        self.fibs.push(Fib::Show { text });
     }
 
     fn cd(&mut self, dir: &str) {
-        self.tale.push(Fib::Run {
+        self.fibs.push(Fib::Run {
             cmd: format!("cd {dir}"),
             result: vec![],
         });
-        self.tale.push(Fib::Look {
+        self.fibs.push(Fib::Look {
             cwd: Some(dir.into()),
             host: None,
             user: None,
@@ -291,7 +287,7 @@ impl LieBuilder {
 
         let apparent_cmd = apparent_cmd.map(ToOwned::to_owned);
         let cmd = cmd.into();
-        self.tale.push(Fib::System { apparent_cmd, cmd });
+        self.fibs.push(Fib::System { apparent_cmd, cmd });
 
         Ok(())
     }
@@ -314,12 +310,12 @@ impl LieBuilder {
             return Err(Box::new(MendaxError::NestedScreens.into()));
         }
 
-        let child = SharedLieBuilder::from(self.child());
+        let child = SharedLie::from(self.child());
         f.call_within_context(&ctx, (child.clone(),))?;
 
-        self.tale.push(Fib::Screen {
+        self.fibs.push(Fib::Screen {
             apparent_cmd: apparent_cmd.map(ToOwned::to_owned),
-            tale: child.lie(&ctx)?.tale.clone(),
+            fibs: child.lie(&ctx)?.fibs.clone(),
         });
 
         Ok(())
@@ -394,7 +390,7 @@ impl LieBuilder {
             }
         }
 
-        self.tale.push(Fib::Look {
+        self.fibs.push(Fib::Look {
             speed,
             title,
             cwd,
@@ -423,16 +419,16 @@ impl LieBuilder {
             return Err(Box::new(MendaxError::DuplicateTag { name }.into()));
         }
 
-        self.tale.push(Fib::Tag { name });
+        self.fibs.push(Fib::Tag { name });
         Ok(())
     }
 
     fn clear(&mut self) {
-        self.tale.push(Fib::Clear);
+        self.fibs.push(Fib::Clear);
     }
 }
 
-impl CustomType for SharedLieBuilder {
+impl CustomType for SharedLie {
     fn build(mut builder: TypeBuilder<Self>) {
         builder
             .with_name("Lie")
@@ -448,6 +444,13 @@ impl CustomType for SharedLieBuilder {
             .with_fn("look", Self::look)
             .with_fn("tag", Self::tag)
             .with_fn("clear", Self::clear);
+    }
+}
+
+#[cfg(test)]
+impl Lie {
+    pub fn fibs(&self) -> &[Fib] {
+        &self.fibs
     }
 }
 
@@ -497,7 +500,7 @@ pub(crate) mod test {
         )?;
 
         assert_eq!(
-            lie.tale().fibs(),
+            lie.fibs(),
             &[
                 Fib::Run {
                     cmd: "foo".into(),
@@ -527,7 +530,7 @@ pub(crate) mod test {
         )?;
 
         assert_eq!(
-            lie.tale().fibs(),
+            lie.fibs(),
             &[Fib::Show {
                 text: "foobar".into()
             }]
@@ -546,7 +549,7 @@ pub(crate) mod test {
         )?;
 
         assert_eq!(
-            lie.tale().fibs(),
+            lie.fibs(),
             &[
                 Fib::Run {
                     cmd: "cd /foo/bar".into(),
@@ -578,7 +581,7 @@ pub(crate) mod test {
             )?;
 
             assert_eq!(
-                lie.tale().fibs(),
+                lie.fibs(),
                 &[
                     Fib::System {
                         cmd: "ls".into(),
@@ -617,14 +620,13 @@ pub(crate) mod test {
                         });
                     "#,
                 )?
-                .tale()
                 .fibs(),
                 &[Fib::Screen {
                     apparent_cmd: None,
-                    tale: Tale::new(vec![Fib::System {
+                    fibs: vec![Fib::System {
                         cmd: "sudo ls /root".into(),
                         apparent_cmd: Some("ls".into()),
-                    }])
+                    }]
                 }]
             );
 
@@ -653,14 +655,13 @@ pub(crate) mod test {
                         });
                     "#,
                 )?
-                .tale()
                 .fibs(),
                 &[Fib::Screen {
                     apparent_cmd: Some("man foo".into()),
-                    tale: Tale::new(vec![Fib::System {
+                    fibs: vec![Fib::System {
                         cmd: "sudo ls /root".into(),
                         apparent_cmd: Some("ls".into()),
-                    }])
+                    }]
                 }]
             );
 
@@ -685,10 +686,7 @@ pub(crate) mod test {
     #[test]
     fn tag() -> Result<(), Box<dyn Error>> {
         assert_eq!(
-            test_script(false, r#"lie.tag("foo")"#)
-                .unwrap()
-                .tale()
-                .fibs(),
+            test_script(false, r#"lie.tag("foo")"#).unwrap().fibs(),
             &[Fib::Tag { name: "foo".into() }],
         );
 
@@ -741,7 +739,7 @@ pub(crate) mod test {
     #[test]
     fn look() -> Result<(), Box<dyn Error>> {
         assert_eq!(
-            test_script(false, r#"lie.look(#{});"#)?.tale().fibs(),
+            test_script(false, r#"lie.look(#{});"#)?.fibs(),
             &[Fib::Look {
                 speed: None,
                 title: None,
@@ -766,7 +764,6 @@ pub(crate) mod test {
                     });
                 "#
             )?
-            .tale()
             .fibs(),
             &[Fib::Look {
                 speed: Some(100.0),
@@ -783,10 +780,7 @@ pub(crate) mod test {
 
     #[test]
     fn clear() -> Result<(), Box<dyn Error>> {
-        assert_eq!(
-            test_script(false, r#"lie.clear()"#)?.tale().fibs(),
-            &[Fib::Clear]
-        );
+        assert_eq!(test_script(false, r#"lie.clear()"#)?.fibs(), &[Fib::Clear]);
 
         Ok(())
     }
